Add flux calculation to poisson.c

+++ src/compute/codes.c
@@ -22,7 +22,7 @@ mr_code_map *mr_code_map_create_from_ocforest(mr_ocforest *forest) {
     map->codes = xmalloc(nb_leaves * sizeof(mr_int));
 
     code_map_userdata ud = { 0, map->codes };
-    for (mr_index octree_idx = 0; octree_idx < forest->nb_roots; ++octree_idx) {
+    for (mr_index octree_idx = 0; (size_t)octree_idx < forest->nb_roots; ++octree_idx) {
         mr_octree_leaves_apply(forest, octree_idx, mr_octree_cond_cb_null(), mr_octree_apply_cb_create(save_code, &ud), false);
     }
 
@@ -39,7 +39,7 @@ void mr_code_map_destroy(mr_code_map *map) {
 }
 
 mr_int mr_code_map_get_code(mr_code_map *map, mr_int idx) {
-    return map && idx < map->len ? map->codes[idx] : MR_INVALID_INDEX;
+    return map && (size_t)idx < map->len ? map->codes[idx] : MR_INVALID_INDEX;
 }
 
 mr_int mr_code_map_get_index(mr_code_map *map, mr_int code) {
